+++ react/components/admin/commissions/ImportCommissionsTab.tsx
@@ -1,10 +1,17 @@
+import { Set } from '@vtex/admin-ui'
 import type { FC } from 'react'
 import React from 'react'
 
 import ImportDropzone from './ImportDropzone'
+import LastImportInfoCard from './LastImportInfoCard'
 
 const ImportCommissionsTab: FC = () => {
-  return <ImportDropzone />
+  return (
+    <Set orientation="vertical" spacing={3} fluid>
+      <ImportDropzone />
+      <LastImportInfoCard />
+    </Set>
+  )
 }
 
 export default ImportCommissionsTab