+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Last imported commission file information and download link
+
 ## [0.42.0] - 2022-03-08
 
 ### Added